Very large commit, added crud for users songs, added refresh and access tokens, fixed many bugs

+++ spotify-clone-back/src/main/java/org/example/repositories/role/IRoleRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface IRoleRepository extends JpaRepository<RoleEntity,Integer> {
+public interface IRoleRepository extends JpaRepository<RoleEntity,Long> {
     Optional<RoleEntity> findByName(String name);
 }
